ZipService

+++ BookMarketRest/src/main/java/com/example/dto/requestDto/ZipcodeRequestDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ZipcodeRequestDto {
     private String name;
-    private Long cityId;
+    private Integer cityId;
 }
